requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
funcsigs==1.0.2
furl==2.1.0
future==0.17.1
gast==0.3.2
grpcio==1.24.1
# -e git+https://github.com/TMats/gym.git@1f11eeb6b27155d0184cb143855eb4d49b9f6dff#egg=gym
h5py==2.10.0
humanfriendly==4.18
idna==2.8
imageio==2.5.0
ipykernel==5.1.1
ipython==7.6.0
ipython-genutils==0.2.0
jedi==0.14.0
Jinja2==2.10.1
joblib==0.14.0
json5==0.8.4
jsonmodels==2.4
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.5.0
jupyterlab==1.0.0
jupyterlab-server==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
mujoco-py==0.5.7
natsort==6.0.0
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
orderedmultidict==1.0.1
pandocfilters==1.4.2
parso==0.5.0
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
pigar==0.9.2
Pillow==6.0.0
plotly==4.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
psutil==5.6.3
ptyprocess==0.6.0
pyglet==1.3.2
Pygments==2.4.2
pyhocon==0.3.51
PyJWT==1.7.1
PyOpenGL==3.1.0
pyparsing==2.4.2
pyrsistent==0.15.2
python-dateutil==2.8.0
PyYAML==5.1.2
pyzmq==18.0.2
requests==2.22.0
requests-file==1.4.3
retrying==1.3.3
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.36.1
trains==0.11.3
traitlets==4.3.2
typing==3.7.4.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0